requirements.txt
Flask==2.3.2
tensorflow==2.12.0
keras==2.12.0
numpy==1.24.3
matplotlib==3.7.1
scikit-learn==1.3.0
Pillow==9.5.0
Werkzeug==2.3.6
